Completed 'adjustClustering' method in DStream; completed Javadoc; updated README.
The 'adjustClustering' method now implements the procedure given in figure 4 of Chen and Tu 2007.
Methods and variables in both CharacteristicVector and GridCluster were updated to generate Javadoc.
README.md was updated to reflect current status and to add the task of implementing a test for connectedness/ability to split clusters.