Wrote DensityGrid.java; Updates to DStream.java and GridCluster.java; Update README.md
Wrote DensityGrid class to represent the density grids used by D-Stream, instead of int[] as previously. The equals and hashCode methods from Object are overwritten to allow the hash maps use to equate Density Grids with the same coordinates.
Updated DStream and GridCluster to reflect this change from int[] to DensityGrids as the keys in various hash maps.
Updated README.md to state that compatability is with MOA 16.04.