Implemented functionality to split  clusters; completed to-do list.
Implemented the isConnected method in GridCluster to test whether a cluster is a valid grid group (Definition 3.4 from Chen and Tu 2007).
Implemented the reCluster method in DStream to recluster an argument GridCluster which has been identified as no longer connected.
Updated the README.md to reflect the completion of the to-do list. Future tasks will be created as issues.