Changes made for correctness.
Added the densityTimeStamp attribute in CharacteristicVector. This ensures that the CharacteristicVector can track both the last time the density was updated (densityTimeStamp) and the last time a record was added to the grid (updateTime).
Went through DStream to ensure that each call to updateTime was correct; changed those to densityTimeStamp that required it.
Cleaned up the isInside() methods for GridCluster. Both return the appropriate boolean value when able rather than maintaining an internal variable.