Update processing for numeric dimensions.
Change to the way in which D-Stream calculates the "size" of a dimension for a numeric attribute. The size now includes one grid "above" and one grid "below" the observed range.

This is done to increase the robustness of the algorithm. It is still recommended to preprocess any numeric data into nominal data as is done in Chen and Tu 2007.