Added 'grid width' parameter, microclustering results  and the ability to calculate inclusion probability.
Dstream and DensityGrid were modified to allow the user to explicitly set the grid width for numeric dimensions. The data preprocessing step was also adapted to this new change. DensityGrids now track their own width.
Dstream was modified to return the DensityGrids as the microclustering result.
DensityGrid and GridCluster were modified to allow for non 0/1 inclusion probabilities based on a point's distance from the nearest DensityGrid in the GridCluster.